% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.legacy_function}
\alias{as.legacy_function}
\title{Tag a function as a legacy mungebit function.}
\usage{
as.legacy_function(x)
}
\arguments{
\item{x}{function. An R function to tag.}
}
\value{
\code{x} with additional class "legacy_mungebit_function".
}
\description{
Tag a function as a legacy mungebit function.
}

