% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{Generic debugging.}
\usage{
debug(fun, text = "", condition = NULL)
}
\arguments{
\item{fun}{any interpreted \R function.}

\item{text}{a text string that can be retrieved when the browser is entered.}

\item{condition}{a condition that can be retrieved when the browser is entered.}
}
\description{
Generic debugging.
}
\seealso{
\code{\link[base]{debug}}
}

