% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standard_column_format.R
\name{except}
\alias{except}
\title{Ignore during standard column format.}
\usage{
except(x)
}
\arguments{
\item{x}{ANY. An R object.}
}
\description{
Ignore during standard column format.
}

