% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.legacy_mungebit}
\alias{is.legacy_mungebit}
\title{Whether a mungebit is a legacy mungebit (from the mungebits package).}
\usage{
is.legacy_mungebit(x)
}
\arguments{
\item{x}{ANY. An R object to test.}
}
\value{
TRUE or FALSE according as the mungebit is a legacy mungebit.
}
\description{
Whether a mungebit is a legacy mungebit (from the mungebits package).
}

