% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.legacy_mungebit_function}
\alias{is.legacy_mungebit_function}
\title{Whether a train or predict function is a legacy function (from the mungebits package).}
\usage{
is.legacy_mungebit_function(x)
}
\arguments{
\item{x}{ANY. An R object to test.}
}
\value{
TRUE or FALSE according as the mungebit is a legacy train or
   predict function, determined using the \code{"legacy_mungebit_function"}
}
\description{
Note that only functions constructed by the \code{munge} helper
will be identifiable using this method.
}

