% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.legacy_mungepiece}
\alias{is.legacy_mungepiece}
\title{Whether a mungepiece is a legacy mungepiece (from the mungepieces package).}
\usage{
is.legacy_mungepiece(x)
}
\arguments{
\item{x}{ANY. An R object to test.}
}
\value{
TRUE or FALSE according as the mungepiece is a legacy mungepiece.
}
\description{
Whether a mungepiece is a legacy mungepiece (from the mungepieces package).
}

