% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mungebit.R
\name{is.mungebit}
\alias{is.mungebit}
\title{Determine whether an object is a mungebit.}
\usage{
is.mungebit(x)
}
\arguments{
\item{x}{ANY. An R object to check.}
}
\value{
TRUE or FALSE according as it has class mungebit.
}
\description{
Determine whether an object is a mungebit.
}
\keyword{typecheck}

