% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungepiece.R
\name{is.mungepiece}
\alias{is.mungepiece}
\title{Determine whether an object is a mungepiece.}
\usage{
is.mungepiece(x)
}
\arguments{
\item{x}{ANY. An R object to check.}
}
\value{
TRUE or FALSE according as it has class mungepiece
}
\description{
Determine whether an object is a mungepiece.
}
\keyword{typecheck}

