% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.transformation}
\alias{is.transformation}
\title{Determine whether or not a given object is a transformation.}
\usage{
is.transformation(x)
}
\arguments{
\item{x}{ANY. R object to test.}
}
\value{
\code{TRUE} or \code{FALSE} according as it is a transformation.
}
\description{
Transformations can be either column or multi column transformations.
}

