% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_column_transformation.R
\name{multi_column_transformation}
\alias{multi_column_transformation}
\title{Multi column transform.}
\usage{
multi_column_transformation()
}
\description{
Multi column transform.
}

