% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/mungebits-package.R
\docType{package}
\name{mungebits2}
\alias{mungebits2}
\alias{mungebits2-package}
\title{Mungebits2 defines a way of thinking about data preparation that
couples the definition of what happens in batch processing versus
online prediction so that both can be described by the same codebase.}
\description{
Mungebits2 defines a way of thinking about data preparation that
couples the definition of what happens in batch processing versus
online prediction so that both can be described by the same codebase.
}

