% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mungepiece-run.R
\docType{data}
\name{mungepiece_run}
\alias{mungepiece_run}
\title{Run a mungepiece and prepare it for a live production setting.}
\format{\preformatted{ NULL
}}
\usage{
mungepiece_run
}
\value{
If the \code{data} parameter is an environment, the transformed
   environment (i.e., the transformed data in the environment) after
   application of the underlying mungebit. If \code{data} is a data.frame,
   the transformed data.frame is returned.
}
\description{
Running a mungepiece achieves the same effect as running the mungebit
attached to the mungepiece: the first time it is run, we \emph{train}
the mungebit so it remembers metadata it will need to replicate the
operation in a live production setting on a single row of data. The
second and subsequent times we run the mungepiece, it will execute
the predict function of the underlying mungebit.
}
\keyword{datasets}

