% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/debug.R
\name{undebug}
\alias{undebug}
\title{Generic removal of debugging.}
\usage{
undebug(fun)
}
\arguments{
\item{fun}{any interpreted \R function.}
}
\description{
Generic removal of debugging.
}
\seealso{
\code{\link[base]{undebug}}
}

