% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/column_transformation.R
\name{column_transformation}
\alias{column_transformation}
\title{Column transform.}
\usage{
column_transformation(...)
}
\arguments{
\item{...}{Random.}
}
\description{
Column transform.
}

