% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mungepiece.R
\name{is.mungepiece}
\alias{is.mungepiece}
\title{Determine whether an object is a mungepiece.}
\usage{
is.mungepiece(x)
}
\description{
Determine whether an object is a mungepiece.
}
\keyword{typecheck}

