% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mungepiece.R
\docType{class}
\name{mungepiece}
\alias{mungepiece}
\title{Mungepiece.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f9a253c1d88> 
 - attr(*, "name")= chr "mungepiece_generator"
}}
\usage{
mungepiece
}
\description{
Mungepiece.
}
\keyword{datasets}

